# docs/swagger.yaml
basePath: /api
definitions:
  handler.Token:
    properties:
      token:
        type: string
    type: object
  handler.email:
    properties:
      email:
        type: string
    type: object
  handler.exerciseDate:
    properties:
      exercise_date:
        type: string
      exercise_time:
        type: integer
    type: object
  handler.exerciseDateForCreate:
    properties:
      exercise_dates:
        items:
          type: string
        type: array
    type: object
  handler.exerciseDateForList:
    properties:
      exercise_dates:
        items:
          $ref: '#/definitions/handler.exerciseDate'
        type: array
      goal:
        type: string
    type: object
  handler.exerciseDateForUpdate:
    properties:
      exercise_dates:
        items:
          $ref: '#/definitions/handler.exerciseDate'
        type: array
    type: object
  handler.goal:
    properties:
      goal:
        type: string
    type: object
  handler.message:
    properties:
      message:
        type: string
    type: object
  handler.onboardingStatus:
    properties:
      is_onboarding_completed:
        type: boolean
    type: object
  handler.signUpUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.Auth:
    properties:
      access_token:
        type: string
      expired_at:
        type: string
    type: object
host: ec2-13-209-98-22.ap-northeast-2.compute.amazonaws.com
info:
  contact: {}
  title: DailyChaCha Sample Swagger API
  version: "1.0"
paths:
  /apple-sign-in:
    post:
      consumes:
      - application/json
      description: Token을 받아 access token을 반환합니다.
      parameters:
      - description: 애플로그인 token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/handler.Token'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Auth'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.message'
      summary: 애플 로그인 API
  /getlist:
    get:
      consumes:
      - application/json
      description: access token을 확인하여 해당 토큰 유저의 이메일을 반환합니다.
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.email'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.message'
      security:
      - ApiKeyAuth: []
      summary: auth 토큰으로 테스트해볼 API입니다.
  /onboarding/dates:
    get:
      consumes:
      - application/json
      description: 유저의 온보딩 여부를 반환합니다.
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.exerciseDateForList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.message'
      security:
      - ApiKeyAuth: []
      summary: 온보딩 운동일정 가져오는 API
    post:
      consumes:
      - application/json
      description: 유저의 온보딩 두번째 - 날짜를 생성하는 API입니다.
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 요일
        in: body
        name: Weekday
        required: true
        schema:
          $ref: '#/definitions/handler.exerciseDateForCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.message'
      security:
      - ApiKeyAuth: []
      summary: 날짜정하기 생성 API
    put:
      consumes:
      - application/json
      description: 유저의 온보딩 세번째 - 시간을 생성하는 API입니다.
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 습관 일정
        in: body
        name: ExerciseDate
        required: true
        schema:
          $ref: '#/definitions/handler.exerciseDateForUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.message'
      security:
      - ApiKeyAuth: []
      summary: 시간정하기 API
  /onboarding/goals:
    get:
      consumes:
      - application/json
      description: 결심하기에서 사용할 목록들을 반환합니다.
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handler.goal'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.message'
      security:
      - ApiKeyAuth: []
      summary: 결심하기 목록 API
    post:
      consumes:
      - application/json
      description: 유저의 온보딩 첫번째 - 결심을 생성하는 API입니다.
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 결심
        in: body
        name: Goal
        required: true
        schema:
          $ref: '#/definitions/handler.goal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.message'
      security:
      - ApiKeyAuth: []
      summary: 결심하기 생성 API
  /onboarding/status:
    get:
      consumes:
      - application/json
      description: 유저의 온보딩 여부를 반환합니다.
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.onboardingStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.message'
      security:
      - ApiKeyAuth: []
      summary: 온보딩 상태 API
  /sign-in:
    post:
      consumes:
      - application/json
      description: email, password를 받아 access token을 반환합니다.
      parameters:
      - description: 유저 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.signUpUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Auth'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.message'
      summary: 로그인 API
  /sign-up:
    post:
      consumes:
      - application/json
      description: email, password를 받아 가입합니다.
      parameters:
      - description: 유저 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.signUpUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.message'
      summary: 회원가입 API
swagger: "2.0"